move document processor router to parsers module
