+++ docs/how-to/document_search/search_documents.md
@@ -76,7 +76,7 @@ library that supports parsing and chunking of most common document types (i.e. p
 from ragbits.core.embeddings.litellm import LiteLLMEmbedder
 from ragbits.core.vector_stores.in_memory import InMemoryVectorStore
 from ragbits.document_search import DocumentSearch
-from ragbits.document_search.ingestion.document_processor import DocumentProcessorRouter
+from ragbits.document_search.ingestion.parsers.router import DocumentProcessorRouter
 from ragbits.document_search.documents.document import DocumentType
 from ragbits.document_search.ingestion.parsers.unstructured.default import UnstructuredDefaultProvider
