+++ packages/ragbits-document-search/tests/integration/test_unstructured.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import pytest
 
 from ragbits.document_search.documents.document import DocumentMeta, DocumentType
-from ragbits.document_search.ingestion.document_processor import DocumentProcessorRouter, ProvidersConfig
+from ragbits.document_search.ingestion.parsers.router import DocumentProcessorRouter, ProvidersConfig
 from ragbits.document_search.ingestion.parsers.unstructured.default import (
     DEFAULT_PARTITION_KWARGS,
     UNSTRUCTURED_API_KEY_ENV,
