+++ packages/ragbits-document-search/tests/unit/test_document_search.py
@@ -18,7 +18,7 @@ from ragbits.document_search.documents.document import (
 )
 from ragbits.document_search.documents.element import TextElement
 from ragbits.document_search.documents.sources import GCSSource, LocalFileSource
-from ragbits.document_search.ingestion.document_processor import DocumentProcessorRouter
+from ragbits.document_search.ingestion.parsers.router import DocumentProcessorRouter
 from ragbits.document_search.ingestion.parsers import BaseProvider
 from ragbits.document_search.ingestion.parsers.dummy import DummyProvider
 from ragbits.document_search.ingestion.strategies.batched import (
